+++ src/main/java/com/mraof/minestuck/data/MinestuckEnUsLanguageProvider.java
@@ -1336,7 +1336,7 @@ public class MinestuckEnUsLanguageProvider extends MinestuckLanguageProvider
 		addItem(MSItems.PLUSH_NAKAGATOR, "Nakagator Plushie");
 		addItem(MSItems.PLUSH_IGUANA, "Iguana Plushie");
 		addItem(MSItems.PLUSH_TURTLE, "Turtle Plushie");
-		addItem(MSItems.PLUSH_MUTINI, "Mutini Plushie");
+		addItem(MSItems.PLUSH_MUTATED_CAT, "Mutated Cat Plushie");
 		addItem(MSItems.IRON_BOAT, "Iron Boat");
 		addItem(MSItems.GOLD_BOAT, "Golden Boat");
 		addItem(MSItems.FLARP_MANUAL, "Flarp Manual");